requirements.txt
django==4.0.6
python-dotenv==0.20.0
python-telegram-bot==13.13